+++ lib/listenable_tools.dart
@@ -1,4 +1,5 @@
 library listenable_tools;
 
+export 'src/async_state.dart';
 export 'src/listenable_widgets.dart';
 export 'src/valuelistenable_widgets.dart';