+++ CHANGELOG.md
@@ -1,3 +1,8 @@
-## 0.0.1
+## 0.0.3
 
-* TODO: Describe initial release.
+* Add AsyncState, AsyncEvent and AsyncNotifier
+* Change All Listenable name by Notifier
+
+## 0.0.2
+
+* Change ListenableBuilder to Listenablebuilder